Speed up explicit 4X by changes to DBC
Profiling revealed that the evaluation of symbolic
expressions for DBCs was comparatively expensive
for explicit time integration.

Refactor DBC to parse the expressions at initialization
and cache the expressions as Julia functions.

Need to propagate these changes to other BCs.

Now bottleneck for explicit integration is the calculation
of the stable time step, which has not been threaded yet.