Checking in input files for bolted-joint problem, non-overlapping.
Currently can only run with tet4s, not tet10s for NO-SAM, as barycentric
coords and interpolation schemes for tri6 element do not exist (see
issue #147, fixing this is WIP).

Solution looks reasonable.  Need to compare with O-SAM for tet4-hex
coupling.  Performance and whether or not run completes is highly
sensitive to the relaxation parameter.  A relaxation parameter of 0.2
seems to give best performance for current (loose) tolerances of 1e-3
and 1e-5, leading to convergence in 13.56 Schwarz iterations on average,
though some steps hit the max number of 32 iters.  O-SAM took 3 Schwarz
iters to converge to tighter tolerance with tet10-hex coupling.

When I try to tighten the tolerance for this problem, the Schwarz method
stagnates (|U| oscillates) and the max # Schwarz iterations are hit for pretty much every
dt.  Perhaps this can be mitigated through better relaxation strategy
like Aitken acceleration, or a more thorough study to identify the best
relaxation parameter.