Corrections to paths for bolted joint problem.
Adding NO-SAM tension specimen problem.  Solution is not correct and
Schwarz behavior is unexpected.  NO-SAM converges in just 2 Schwarz
iterations / dt, vs. up to 32+ for O-SAM for this problem.  Solution is
clearly not being transferred to bottom subdomain.  I checked the
notched cylinder results I generated awhile ago, which is similar, and
it looks like I could never get NO-SAM to work there, although SAM was
taking more iterations for that problem.  Other similar pulling problems
did work with NO-SAM, like the cylinder and cuboid.  Not sure what is
going on with the tension specimen.