The original NO bolted joint problem I created, with the bolt having its own separate domain, effectively assumed there was something like rust holding things together around the bold. This is a different setup than the O-SAM version of this problem.
I have now created a version of the NO bolted joint problem in
which the bolts domain has a small part of the parts and bottom plate
as a part of it, which should give behavior closer to the O-SAM verison
of this problem.  Still needs to be run/tested.